Renamed Parallel to JROMP and updated javadoc
