+++ src/test/java/jromp/construct/atomic/AtomicTests.java
@@ -1,6 +1,6 @@
 package jromp.construct.atomic;
 
-import jromp.Parallel;
+import jromp.JROMP;
 import jromp.operation.Operations;
 import jromp.var.SharedVariable;
 import jromp.var.Variables;
@@ -13,15 +13,15 @@ class AtomicTests {
     void testRead() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                    Integer value = Atomic.read("x", vars);
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Integer value = Atomic.read("x", vars);
 
-                    assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
-                    assertThat(value).isZero();
-                })
-                .join();
+                 assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
+                 assertThat(value).isZero();
+             })
+             .join();
 
         // The atomic variable is removed after the parallel block is executed
         assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isFalse();
@@ -31,16 +31,16 @@ class AtomicTests {
     void testWrite() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                    Atomic.write("x", 1, vars);
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Atomic.write("x", 1, vars);
 
-                    // Only the atomic variable should be updated here
-                    assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
-                    assertThat(variables.<Integer>get("x").value()).isZero();
-                })
-                .join();
+                 // Only the atomic variable should be updated here
+                 assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
+                 assertThat(variables.<Integer>get("x").value()).isZero();
+             })
+             .join();
 
         // The atomic variable changes its value to the final value on the "parent" shared variable
         // after the parallel block is executed, and the atomic variable is removed.
@@ -52,16 +52,16 @@ class AtomicTests {
     void testUpdate() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                    Atomic.update("x", Operations.add(1), vars);
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Atomic.update("x", Operations.add(1), vars);
 
-                    // Only the atomic variable should be updated here
-                    assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
-                    assertThat(variables.<Integer>get("x").value()).isZero();
-                })
-                .join();
+                 // Only the atomic variable should be updated here
+                 assertThat(((SharedVariable<Integer>) variables.<Integer>get("x")).hasAtomic()).isTrue();
+                 assertThat(variables.<Integer>get("x").value()).isZero();
+             })
+             .join();
 
         // The atomic variable changes its value to the final value on the "parent" shared variable
         // after the parallel block is executed, and the atomic variable is removed.
