+++ src/test/java/jromp/construct/critical/CriticalTests.java
@@ -1,6 +1,6 @@
 package jromp.construct.critical;
 
-import jromp.Parallel;
+import jromp.JROMP;
 import jromp.var.SharedVariable;
 import jromp.var.Variable;
 import jromp.var.Variables;
@@ -13,10 +13,10 @@ class CriticalTests {
     void testSimple() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1)))
-                .join();
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1)))
+             .join();
 
         assertThat(variables.get("x").value()).isEqualTo(4);
     }
@@ -25,13 +25,13 @@ class CriticalTests {
     void testSameNameMulti() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                           Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
-                           Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
-                       })
-                .join();
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
+                 Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
+             })
+             .join();
 
         assertThat(variables.get("x").value()).isEqualTo(8);
     }
@@ -42,13 +42,13 @@ class CriticalTests {
                                        .add("x", new SharedVariable<>(0))
                                        .add("y", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                           Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
-                           Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
-                       })
-                .join();
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
+                 Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
+             })
+             .join();
 
         assertThat(variables.get("x").value()).isEqualTo(4);
         assertThat(variables.get("y").value()).isEqualTo(4);
@@ -60,15 +60,15 @@ class CriticalTests {
                                        .add("x", new SharedVariable<>(0))
                                        .add("y", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                           Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
-                           Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
-                           Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
-                           Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
-                       })
-                .join();
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
+                 Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
+                 Critical.enter("x", id, vars, (i, v) -> v.<Integer>get("x").update(x -> x + 1));
+                 Critical.enter("y", id, vars, (i, v) -> v.<Integer>get("y").update(y -> y + 1));
+             })
+             .join();
 
         assertThat(variables.get("x").value()).isEqualTo(8);
         assertThat(variables.get("y").value()).isEqualTo(8);
@@ -78,17 +78,17 @@ class CriticalTests {
     void testMoreOperationsInsideCritical() {
         Variables variables = Variables.create().add("x", new SharedVariable<>(new StringBuilder()));
 
-        Parallel.withThreads(4)
-                .withVariables(variables)
-                .block((id, vars) -> {
-                           Critical.enter("x", id, vars, (i, v) -> {
-                               Variable<StringBuilder> variable = v.get("x");
-                               variable.update(sb -> sb.append("x"));
-                               variable.update(sb -> sb.append(" "));
-                               variable.update(sb -> sb.append("more"));
-                           });
-                       })
-                .join();
+        JROMP.withThreads(4)
+             .withVariables(variables)
+             .block((id, vars) -> {
+                 Critical.enter("x", id, vars, (i, v) -> {
+                     Variable<StringBuilder> variable = v.get("x");
+                     variable.update(sb -> sb.append("x"));
+                     variable.update(sb -> sb.append(" "));
+                     variable.update(sb -> sb.append("more"));
+                 });
+             })
+             .join();
 
         assertThat(variables.get("x").value().toString()).hasToString("x more".repeat(4));
     }
