+++ src/test/java/jromp/var/AtomicVariableTests.java
@@ -1,6 +1,6 @@
 package jromp.var;
 
-import jromp.Parallel;
+import jromp.JROMP;
 import jromp.utils.Utils;
 import org.junit.jupiter.api.Test;
 
@@ -14,16 +14,16 @@ class AtomicVariableTests {
         int[] countsPerThread = new int[threads];
         Variables vars = Variables.create().add("sum", new AtomicVariable<>(0));
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        Variable<Integer> insideSum = variables.get("sum");
-                        insideSum.update(old -> old + 1);
-                        countsPerThread[id]++;
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     Variable<Integer> insideSum = variables.get("sum");
+                     insideSum.update(old -> old + 1);
+                     countsPerThread[id]++;
+                 }
+             })
+             .join();
 
         assertThat(vars.get("sum").value()).isEqualTo(iterations);
         assertThat(countsPerThread).containsOnly(iterations / threads);
@@ -37,15 +37,15 @@ class AtomicVariableTests {
         AtomicVariable<Integer> outsideSum = new AtomicVariable<>(0);
         Variables vars = Variables.create().add("sum", outsideSum);
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        outsideSum.update(old -> old + 1);
-                        countsPerThread[id]++;
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     outsideSum.update(old -> old + 1);
+                     countsPerThread[id]++;
+                 }
+             })
+             .join();
 
         assertThat(outsideSum.value()).isEqualTo(iterations);
         assertThat(countsPerThread).containsOnly(iterations / threads);
@@ -57,15 +57,15 @@ class AtomicVariableTests {
         int iterations = 10;
         Variables vars = Variables.create().add("sum", new AtomicVariable<>(0));
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        Variable<Integer> sum = variables.get("sum");
-                        sum.set(1);
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     Variable<Integer> sum = variables.get("sum");
+                     sum.set(1);
+                 }
+             })
+             .join();
 
         assertThat(vars.get("sum").value()).isEqualTo(1);
     }
@@ -83,27 +83,27 @@ class AtomicVariableTests {
     void testKeepLastValueAfterExecution() {
         Variables vars = Variables.create().add("sum", new AtomicVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(vars)
-                .block((id, variables) -> {
-                    for (int i = 0; i < 2; i++) {
-                        Variable<Integer> sum = variables.get("sum");
+        JROMP.withThreads(4)
+             .withVariables(vars)
+             .block((id, variables) -> {
+                 for (int i = 0; i < 2; i++) {
+                     Variable<Integer> sum = variables.get("sum");
 
-                        // The master thread will sleep for a while to end up with a different value
-                        if (Utils.isMaster(id)) {
-                            try {
-                                Thread.sleep(100);
-                            } catch (InterruptedException e) {
-                                throw new RuntimeException(e);
-                            }
+                     // The master thread will sleep for a while to end up with a different value
+                     if (Utils.isMaster(id)) {
+                         try {
+                             Thread.sleep(100);
+                         } catch (InterruptedException e) {
+                             throw new RuntimeException(e);
+                         }
 
-                            sum.set(200);
-                        } else {
-                            sum.update(old -> old + 1);
-                        }
-                    }
-                })
-                .join();
+                         sum.set(200);
+                     } else {
+                         sum.update(old -> old + 1);
+                     }
+                 }
+             })
+             .join();
 
         assertThat(vars.get("sum").value()).isEqualTo(200);
     }
