+++ src/test/java/jromp/var/SharedVariableTests.java
@@ -1,6 +1,6 @@
 package jromp.var;
 
-import jromp.Parallel;
+import jromp.JROMP;
 import jromp.utils.Utils;
 import org.assertj.core.api.Condition;
 import org.junit.jupiter.api.Test;
@@ -15,16 +15,16 @@ class SharedVariableTests {
         int[] countsPerThread = new int[threads];
         Variables vars = Variables.create().add("sum", new SharedVariable<>(0));
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        Variable<Integer> insideSum = variables.get("sum");
-                        insideSum.update(old -> old + 1);
-                        countsPerThread[id]++;
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     Variable<Integer> insideSum = variables.get("sum");
+                     insideSum.update(old -> old + 1);
+                     countsPerThread[id]++;
+                 }
+             })
+             .join();
 
         // The value of the shared variable should be less than or equal to the number of iterations
         // because the update operation is not atomic and race conditions can occur.
@@ -41,15 +41,15 @@ class SharedVariableTests {
         SharedVariable<Integer> outsideSum = new SharedVariable<>(0);
         Variables vars = Variables.create().add("sum", outsideSum);
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        outsideSum.update(old -> old + 1);
-                        countsPerThread[id]++;
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     outsideSum.update(old -> old + 1);
+                     countsPerThread[id]++;
+                 }
+             })
+             .join();
 
         // The value of the shared variable should be less than or equal to the number of iterations
         // because the update operation is not atomic and race conditions can occur.
@@ -64,15 +64,15 @@ class SharedVariableTests {
         int iterations = 10;
         Variables vars = Variables.create().add("sum", new SharedVariable<>(0));
 
-        Parallel.withThreads(threads)
-                .withVariables(vars)
-                .parallelFor(0, iterations, false, (id, start, end, variables) -> {
-                    for (int i = start; i < end; i++) {
-                        Variable<Integer> sum = variables.get("sum");
-                        sum.set(1);
-                    }
-                })
-                .join();
+        JROMP.withThreads(threads)
+             .withVariables(vars)
+             .parallelFor(0, iterations, false, (id, start, end, variables) -> {
+                 for (int i = start; i < end; i++) {
+                     Variable<Integer> sum = variables.get("sum");
+                     sum.set(1);
+                 }
+             })
+             .join();
 
         assertThat(vars.get("sum").value()).isEqualTo(1);
     }
@@ -90,27 +90,27 @@ class SharedVariableTests {
     void testKeepLastValueAfterExecution() {
         Variables vars = Variables.create().add("sum", new SharedVariable<>(0));
 
-        Parallel.withThreads(4)
-                .withVariables(vars)
-                .block((id, variables) -> {
-                    for (int i = 0; i < 2; i++) {
-                        Variable<Integer> sum = variables.get("sum");
+        JROMP.withThreads(4)
+             .withVariables(vars)
+             .block((id, variables) -> {
+                 for (int i = 0; i < 2; i++) {
+                     Variable<Integer> sum = variables.get("sum");
 
-                        // The master thread will sleep for a while to end up with a different value
-                        if (Utils.isMaster(id)) {
-                            try {
-                                Thread.sleep(100);
-                            } catch (InterruptedException e) {
-                                throw new RuntimeException(e);
-                            }
+                     // The master thread will sleep for a while to end up with a different value
+                     if (Utils.isMaster(id)) {
+                         try {
+                             Thread.sleep(100);
+                         } catch (InterruptedException e) {
+                             throw new RuntimeException(e);
+                         }
 
-                            sum.set(200);
-                        } else {
-                            sum.update(old -> old + 1);
-                        }
-                    }
-                })
-                .join();
+                         sum.set(200);
+                     } else {
+                         sum.update(old -> old + 1);
+                     }
+                 }
+             })
+             .join();
 
         assertThat(vars.get("sum").value()).isEqualTo(200);
     }
